Initial work to generate a score card for a threat actor
The simple command-line processing has been upgraded to use an argument
processing package, because there is a bit more sophistication required
now.

There are some code reorganisations for improved maintainability, such as
moving the chunk of code that enumerates which heatmaps to generate into
heatmaps.py, and the code that identifies the threat actors into a new
module, utility.py.

Currently the score card is generated for a single named actor, but my
current thinking is that it should probably scan the MISP data once and
generate score cards for all actors (rather than scan multiple times to
generate score cards for a number of different actors.)