Some tweaks from investigating possibilities for representing uncertainty/errors
Attempts to show uncertainty have all failed so far due to gnuplot
idiosyncracies (for example, different versions of gnuplot produce
wildly different plots for "image" styles.  Error bars didn't work
out (and won't look great anyway.)

What would be idea is to produce an oversize bar as an image using
a gradient such as:

    # Set the palette
    #
    outfile.write("set palette defined (0.0 \"white\", " +
        str(score_range[score] * 0.5) + " \"white\", " +
        str(score_range[score]) + " \"" + score_colour[score] + "\", " +
        str(score_range[score] * 1.5) + " \"white\")\n")

That ought to be able to produce a nice fuzzy gradient around the
desired score.  However, it seems to be impossible to get all of the
requirements absolutely correct, and if anything is wrong gnuplot
produces a cryptic message.  For example, when specifying variable
colour in the plot, or using a palette, gnuplot won't do anything
unless you add another *two* data items to the plot (where all the
examples I can find show the expected *one* data item being added.)
I can only hazard a guess as to what a headache gnuplot maintenance
must be.