Add initial gradients to scores
The point to note is that gnuplot won't handle certain things when
using labelled data.  For example, if you try to specify a box width
when using labelled data it will just treat the data as spurious (only
warning you some of the time) and will just not draw anything.  I
suppose it makes some sense that if the data isn't numeric then what
does a width signify?

If you want to specify variable colours for boxes, then you have to
specify a box width: "Them's the rules!"  So basically you are forced
into not using labelled data.  But then the xtics need to be hidden,
because they are irrelevant, and xlabels need to be used to label the
scores.  Only, in this situation the xlabels can't be rotated.

Basically, gnuplot is a bit of a mess.  Different versions also
generate wildly different outputs for the same inputs :-/

Also note that only the first palette gets used, for some reason.
To be looked into...