Tweak the scaling so that the format of the scorecards can be reviewed
Also: Apply the logarithm before applying the multiplier so that the
multiplier is easier to intuit.

Again, the scorings are currently arbitrary.  The MISP data seems to be
quite exponential in nature, so fixing scores for the most significant
threat actors results in unrealistically low scores for other threat actors.
This might well be a natural consequence of the way incident data enters MISP.