Improve the behaviour of the caching
The MISP data is now processed as follows:
- The data present in the cache will be read
- All events are downloaded from the server
- Attributes for events which are not present in the cache will be downloaded
- The data is written back to the cache
- The visualisations are generated from the data

Note that this will not download attributes for events that are already
in the cache.  Use the option --forcedownload to force a full download
of attributes from the server, but note that this can take some time.

The code no longer makes an attempt to recover from interruption. If an
interruption does occur during download, then it will be necessary to
manually force a full download of all MISP data.

A future commit may attempt to only cache events for which the corresponding
attributes have been successfully downloaded.  This would allow the full
MISP data to be downloaded in multiple settings.  As it is currently,
permitting interruption of attribute download would mean that all of the
event data would be downloaded but attributes would only be partially
downloaded (and it couldn't be redownloaded without forcing a full
download.)