Rework the argument handling: --heatmaps is required to generate heatmaps
This reworks the argument handling so that you can specify generation of any
or all of: heatmaps, scorecards, and scatter plots.  Heatmap generation is no
longer the default.

It is also possible to perform other actions independently, such as listing
threat actors, analysing the MISP data, or dumping the cache (which now implies
avoiding querying the server.)